+++ migrations/2021-09-29-130958_strategy_override/down.sql
@@ -1,0 +1,1 @@
+ALTER TABLE pull_request DROP COLUMN strategy_override;