+++ migrations/2021-09-29-130958_strategy_override/up.sql
@@ -1,0 +1,1 @@
+ALTER TABLE pull_request ADD COLUMN strategy_override varchar(255);